Fix several bugs in integrators

+++ ABIE/ode.py
@@ -44,7 +44,7 @@ class ODE(object):
 
         # Differential equations:
         # - Position
-        nbodies = x.size / 6
+        nbodies = int(x.size / 6)
         dxdt[0:nbodies*3] = x[nbodies*3:]  # velocities
         for j in range(0, nbodies):
             # dxdt[j * 3:3 + j * 3] = x[nbodies * 3 + j * 3:nbodies * 3 + 3 + j * 3]
@@ -65,7 +65,7 @@ class ODE(object):
     @staticmethod
     def ode_n_body_second_order(x, const_g, masses):
 
-        nbodies = x.size / 3  # WARNING: this x contains only positions!!
+        nbodies = int(x.size / 3)  # WARNING: this x contains only positions!!
 
         # Allocate
         acc = x * 0.0
